chore(flows): cleanup unused code and rename legacy factory files
Cleanup:
- Remove unused AgentLifecycle methods: toSnapshot(), fromSnapshot(), setStatus()
- Remove unused exports: AgentLifecycleSnapshotSchema, AgentLifecycleSnapshot
- Make AGENT_LIFECYCLE_STATUS and AgentLifecycleStatus private (not exported)
- Simplify AgentLifecycle to essential methods only

File renames (factory → class naming):
- createFinalizeNode.ts → StandardFinalizeNode.ts
- createInitNode.ts → StandardInitNode.ts

The snapshot/serialization methods were never called anywhere in the codebase.
The setStatus method was also unused. This reduces the public API surface.
