+++ src/agent/implementations/flows/common/index.ts
@@ -16,5 +16,5 @@
 
 export * from './AgentLifecycle';
 export * from './AgentRunFlowRunner';
-export * from './createFinalizeNode';
-export * from './createInitNode';
+export * from './StandardFinalizeNode';
+export * from './StandardInitNode';
